feat(filter/date-popover): refactor code, split components, etc.
