+++ src/constants/options.js
@@ -1,5 +1,5 @@
 import { PROPERTY_TYPE_ICONS } from './icons'
-import { DATE_RANGE_TEXT, PROPERTY_TYPE_TEXT, TIME_UNIT_TEXT } from './texts'
+import { DATE_RANGE_TEXT, PLURAL_TIME_UNIT_TEXT, PROPERTY_TYPE_TEXT, TIME_UNIT_TEXT } from './texts'
 import { DATE_RANGE_TYPE, GENDER, PROPERTY_TYPE, TIME_UNIT_TYPE } from './types'
 
 export const GENDER_OPTIONS = Object.values(GENDER).map((gender) => ({
@@ -20,5 +20,10 @@ export const DATE_RANGE_OPTIONS = Object.values(DATE_RANGE_TYPE).map((type) => (
 
 export const TIME_UNIT_OPTIONS = Object.values(TIME_UNIT_TYPE).map((type) => ({
   label: TIME_UNIT_TEXT[type],
+  value: type,
+}))
+
+export const PLURAL_TIME_UNIT_OPTIONS = Object.values(TIME_UNIT_TYPE).map((type) => ({
+  label: PLURAL_TIME_UNIT_TEXT[type],
   value: type,
 }))
